Move random classes to separate folder.
