+++ resolve/LinSolverIterativeRandFGMRES.cpp
@@ -7,8 +7,8 @@
 #include <resolve/utilities/logger/Logger.hpp>
 #include <resolve/matrix/MatrixHandler.hpp>
 #include "LinSolverIterativeRandFGMRES.hpp"
-#include <resolve/RandSketchingCountSketch.hpp> 
-#include <resolve/RandSketchingFWHT.hpp> 
+#include <resolve/random/RandSketchingCountSketch.hpp> 
+#include <resolve/random/RandSketchingFWHT.hpp> 
 
 namespace ReSolve
 {
