+++ resolve/random/RandSketchingCountSketch.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <resolve/Common.hpp>
-#include <resolve/RandSketchingManager.hpp>
+#include <resolve/random/RandSketchingManager.hpp>
 
 namespace ReSolve {
