+++ resolve/random/RandSketchingManager.hpp
@@ -1,8 +1,16 @@
 // this is a virtual class
 #pragma once
 #include <resolve/Common.hpp>
-#include <resolve/vector/Vector.hpp>
 #include <resolve/MemoryUtils.hpp>
+
+
+namespace ReSolve
+{
+  namespace vector
+  {
+    class Vector;
+  }
+}
 
 namespace ReSolve { 
   class RandSketchingManager {
@@ -28,9 +36,9 @@ namespace ReSolve {
       virtual index_type getPaddedSize();
 
     protected:
-      index_type n_;///< size of base vector
+      index_type n_;      ///< size of base vector
       index_type k_rand_; ///< size of sketched vector
-      index_type N_; ///< padded n -- generally N_ > n_
+      index_type N_;      ///< padded n -- generally N_ > n_
     
       MemoryHandler mem_; ///< Device memory manager object
   };
